requirements.txt
fastapi
uvicorn
gunicorn
pydantic
google-generativeai